src/requirements.txt
dash>=2.14.0
plotly>=5.15.0
packaging>=21.3
flask>=2.3.3
flask-login>=0.6.3
oauthlib>=3.2.2
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.3